+++ HISTORY.rst
@@ -2,10 +2,15 @@
 History
 =======
 
+0.4.0 (2021-10-12)
+------------------
+
+* Add validate permission feature.
+
 0.3.0 (2021-10-05)
 ------------------
 
-* Add authentication and autorization feature.
+* Add client token grant feature.
 
 0.1.0 (2021-09-09)
 ------------------